Switch package name to "org.physical_web.cms"
