+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 25
     buildToolsVersion "25.0.3"
     defaultConfig {
-        applicationId "com.physical_web.cms.physicalwebcms"
+        applicationId "org.physical_web.cms"
         minSdkVersion 18
         targetSdkVersion 25
         versionCode 1
