--- app/src/main/java/com/physical_web/cms/physicalwebcms/bluetooth/BeaconScanningListener.java
@@ -1,7 +1,0 @@
-package com.physical_web.cms.physicalwebcms.bluetooth;
-
-import android.bluetooth.BluetoothDevice;
-
-public interface BeaconScanningListener {
-    void onBeaconDiscovered(BluetoothDevice beacon);
-}
