--- app/src/main/java/com/physical_web/cms/physicalwebcms/exhibits/Exhibit.java
@@ -1,9 +1,0 @@
-package com.physical_web.cms.physicalwebcms.exhibits;
-
-/**
- * Created by Kayali on 2017-07-12.
- */
-
-public class Exhibit {
-    String title;
-}
