--- app/src/main/java/com/physical_web/cms/physicalwebcms/exhibits/ExhibitFragment.java
@@ -1,43 +1,0 @@
-package com.physical_web.cms.physicalwebcms.exhibits;
-
-import android.app.Fragment;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.physical_web.cms.physicalwebcms.R;
-
-/**
- * Fragment for displaying information about exhibits
- */
-public class ExhibitFragment extends Fragment {
-    public ExhibitFragment() {
-        // Required empty public constructor
-    }
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @return A new instance of fragment ExhibitFragment.
-     */
-    public static ExhibitFragment newInstance() {
-        ExhibitFragment fragment = new ExhibitFragment();
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-        }
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_exhibit, container, false);
-    }
-}
