--- app/src/main/java/com/physical_web/cms/physicalwebcms/sync/SyncStatusListener.java
@@ -1,5 +1,0 @@
-package com.physical_web.cms.physicalwebcms.sync;
-
-public interface SyncStatusListener {
-    void syncStatusChanged(int status);
-}
