+++ app/src/main/java/org/physical_web/cms/bluetooth/BeaconScanningListener.java
@@ -1,0 +1,7 @@
+package org.physical_web.cms.bluetooth;
+
+import android.bluetooth.BluetoothDevice;
+
+public interface BeaconScanningListener {
+    void onBeaconDiscovered(BluetoothDevice beacon);
+}
