+++ app/src/main/java/org/physical_web/cms/exhibits/ImageContent.java
@@ -1,0 +1,8 @@
+package org.physical_web.cms.exhibits;
+
+/**
+ * Created by Kayali on 2017-07-12.
+ */
+
+public class ImageContent extends ExhibitContent {
+}
