+++ app/src/main/java/org/physical_web/cms/sync/SyncStatusListener.java
@@ -1,0 +1,5 @@
+package org.physical_web.cms.sync;
+
+public interface SyncStatusListener {
+    void syncStatusChanged(int status);
+}
