+++ app/src/main/res/layout/activity_drive_setup.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.physical_web.cms.physicalwebcms.setup.DriveSetupActivity"
+        tools:context="org.physical_web.cms.setup.DriveSetupActivity"
         tools:layout_editor_absoluteY="81dp"
         tools:layout_editor_absoluteX="0dp"
