+++ app/src/main/res/layout/activity_enrollment.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <com.flipboard.bottomsheet.BottomSheetLayout android:id="@+id/bottomsheet"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.physical_web.cms.physicalwebcms.beacons.EnrollmentActivity"
+    tools:context="org.physical_web.cms.beacons.EnrollmentActivity"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:layout_editor_absoluteX="0dp"
