+++ app/src/main/res/layout/content_welcome.xml
@@ -5,7 +5,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context="com.physical_web.cms.physicalwebcms.setup.WelcomeActivity"
+        tools:context="org.physical_web.cms.setup.WelcomeActivity"
         tools:showIn="@layout/activity_welcome"
         android:layout_margin="15dp"
         android:layout_weight="0.10"
