+++ app/src/main/res/layout/fragment_about.xml
@@ -2,7 +2,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".AboutFragment"
+    tools:context="org.physical_web.cms.AboutFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="8dp">
