+++ app/src/main/res/layout/fragment_beacon.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".beacons.BeaconFragment">
+    tools:context="org.physical_web.cms.beacons.BeaconFragment">
 
     <com.flipboard.bottomsheet.BottomSheetLayout
         android:id="@+id/beacon_bottomsheet"
