+++ app/src/main/res/layout/fragment_welcome.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WelcomeFragment">
+    tools:context="org.physical_web.cms.WelcomeFragment">
 
     <android.support.v7.widget.CardView
         android:id="@+id/welcome_exhibit_card"
